avoid recreating the same GPU resource when calling AbstractResource::upload() and check that the resource was actually allocated before destroying it in AbstractResource::dispose()
Former-commit-id: d1f18e51b7af379c7c041ed2fd62b98ba488b83f
Former-commit-id: 8366505ccd408a356838cf15a5036229ae4e6a70

+++ framework/src/minko/render/IndexBuffer.cpp
@@ -26,13 +26,17 @@ using namespace minko::render;
 void
 IndexBuffer::upload()
 {
-	_id = _context->createIndexBuffer(_data.size());
+    if (_id == -1)
+    	_id = _context->createIndexBuffer(_data.size());
 	_context->uploaderIndexBufferData(_id, 0, _data.size(), &_data[0]);
 }
 
 void
 IndexBuffer::dispose()
 {
-	_context->deleteIndexBuffer(_id);
-	_id = -1;
+    if (_id != -1)
+    {
+	    _context->deleteIndexBuffer(_id);
+	    _id = -1;
+    }
 }
